+++ src/commands/handlers/ticket/index.ts
@@ -7,4 +7,5 @@ export * from './typeFields';
 export * from './typeList';
 export * from './typeRemove';
 export * from './typeToggle';
+export * from './userRestrict';